REFACTOR: Move model analyses to top-level module.
