+++ packages/frontend/src/analysis/index.ts
