+++ packages/frontend/src/analysis/model_graph.tsx
@@ -1,10 +1,10 @@
 import type * as Viz from "@viz-js/viz";
 
-import type { ModelAnalysisProps, ModelJudgment } from "../../model";
-import type { ModelAnalysisMeta, Theory, TypeMeta } from "../../theory";
-import { GraphvizSVG } from "../../visualization";
+import type { ModelAnalysisProps, ModelJudgment } from "../model";
+import type { ModelAnalysisMeta, Theory, TypeMeta } from "../theory";
+import { GraphvizSVG } from "../visualization";
 
-import styles from "../styles.module.css";
+import styles from "../stdlib/styles.module.css";
 
 /** Configure a graph visualization for use with models of a double theory. */
 export function configureModelGraph(options: {
