+++ packages/frontend/src/analysis/submodel_graphs.css
