+++ packages/frontend/src/analysis/submodel_graphs.tsx
@@ -4,9 +4,9 @@ import ChevronRight from "lucide-solid/icons/chevron-right";
 import { Show } from "solid-js";
 
 import type { DblModel } from "catlog-wasm";
-import { IconButton } from "../../components";
-import type { ModelAnalysisProps, ModelJudgment } from "../../model";
-import type { ModelAnalysisMeta, Theory } from "../../theory";
+import { IconButton } from "../components";
+import type { ModelAnalysisProps, ModelJudgment } from "../model";
+import type { ModelAnalysisMeta, Theory } from "../theory";
 import { type GraphvizAttributes, ModelGraphviz } from "./model_graph";
 
 import "./submodel_graphs.css";
